Convert to DList and Printer monad.
Reorganize the code such that source locations are handled on the printer,
rather than on the parser side. This makes the output more sensible and the
code cleaner (although there is quite a bit of new code).

Syntax trees store the music text coupled with the source locations as DLists.

This commit also adds the feature for adding music to all current parts with
`*` which was already advertised in `README.md`.